requirements.txt
Flask==3.0.0
PyMongo==4.6.1
pydantic==2.5.3
python-dotenv==1.0.0
Faker==22.2.0

# Development dependencies
ruff==0.1.9
black==23.12.1
pytest==7.4.4
pytest-flask==1.3.0